Updated S3Cycle_Sampler
Rewrote function so that there is only an "if" and then "else" statement. The result is that the 3-cycle with a reticulation vertex adjacent to the leaf edge is sampled with probability 2/3, and the 3-cycle with internal reticulation vertex is sampled with probability 1/3.

Changed Tr1 and Tr2 in both the "if" and "else" statements to be compatible with a single rooting of corresponding semi-directed network (along the edge labelled by l5 in the first statement and along the edge labelled by l3 in the second).