Changed Trees in Two3Cycle_Sampler
Edited the Two3Cycle_Sampler to have just an "if" and "else" loop. The "if" loop corresponds to the double 3-cycle with two leaf edge adjacent reticulations which is sampled 2/3 of the time and the "else" loop to the double 3-cycle with one internal reticulation vertex.

Also updated "lengths" so that the interval from which edges 5, 6, 7, and 8 are drawn is 1/3 the max length. These are the only edges that appear in a triple edge in some of the constituent trees.